+++ lib/common/widgets/bottom_sheet.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/material.dart';
+
+class BottomSheetTemplate extends StatelessWidget {
+  final Widget child;
+
+  const BottomSheetTemplate({
+    Key? key,
+    required this.child,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return SafeArea(
+      bottom: true,
+      child: Container(
+        decoration: const BoxDecoration(
+          borderRadius: BorderRadius.only(
+            topLeft: Radius.circular(8),
+            topRight: Radius.circular(8),
+          ),
+        ),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Align(
+              alignment: Alignment.center,
+              child: Padding(
+                padding: const EdgeInsets.symmetric(
+                  vertical: 21,
+                ),
+                child: Container(
+                  height: 4,
+                  width: 70,
+                  decoration: BoxDecoration(
+                    borderRadius: const BorderRadius.all(
+                      Radius.circular(12),
+                    ),
+                    color: Colors.grey.withOpacity(0.6),
+                  ),
+                ),
+              ),
+            ),
+            Expanded(child: child),
+          ],
+        ),
+      ),
+    );
+  }
+}