+++ lib/presentation/controllers/activity_filter_provider.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class ActivityFilterProvider extends ChangeNotifier {
+  List<int?> selectedStatusCodes = [];
+
+  void onChangeSelectedStatusCode(int statusCode) {
+    if (selectedStatusCodes.contains(statusCode)) {
+      selectedStatusCodes.remove(statusCode);
+    } else {
+      selectedStatusCodes.add(statusCode);
+    }
+    notifyListeners();
+  }
+}